% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar.R
\name{import}
\alias{import}
\title{This function imports/loads packages as in Python, i.e., ``import package as alias''}
\usage{
import(package, alias)
}
\arguments{
\item{package}{Package name (unquoted).}

\item{alias}{Alias (unquoted) for the package.}
}
\description{
This function imports/loads packages as in Python, i.e., ``import package as alias''
}
\examples{
import(dplyr, d)
df <- data.frame(a=1:3, b=4:6)
df \%>\% d$filter(a == 2)
}
